enhanced and simplified .bashrc

+++ .bashrc
@@ -87,11 +87,6 @@ fi
 # colored GCC warnings and errors
 #export GCC_COLORS='error=01;31:warning=01;35:note=01;36:caret=01;32:locus=01:quote=01'
 
-# some more ls aliases
-alias ll='ls -alF'
-alias la='ls -A'
-alias l='ls -CF'
-
 # Add an "alert" alias for long running commands.  Use like so:
 #   sleep 10; alert
 alias alert='notify-send --urgency=low -i "$([ $? = 0 ] && echo terminal || echo error)" "$(history|tail -n1|sed -e '\''s/^\s*[0-9]\+\s*//;s/[;&|]\s*alert$//'\'')"'
@@ -123,13 +118,22 @@ export PATH="/opt/nvim/bin:$PATH"
 export NVM_DIR="$([ -z "${XDG_CONFIG_HOME-}" ] && printf %s "${HOME}/.nvm" || printf %s "${XDG_CONFIG_HOME}/nvm")"
 [ -s "$NVM_DIR/nvm.sh" ] && \. "$NVM_DIR/nvm.sh" # This loads nvm
 
-# auto venv
 venv() {
     if [ ! -d "venv" ]; then
         python -m venv venv
+        source venv/bin/activate
+        [ -f "requirements.txt" ] && pip install -r requirements.txt
+    else
+        source venv/bin/activate
     fi
-    source venv/bin/activate
-    if [ -f "requirements.txt" ]; then
-        pip install -r requirements.txt
+}
+
+revenv() {
+    if [[ -n "$VIRTUAL_ENV" ]]; then
+        deactivate
     fi
+    if [ -d "venv" ]; then
+        rm -rf venv
+    fi
+    venv force
 }
